refactor: Board 생성자에서 Interface 를 받아 생성하도록 변경
