+++ src/test/java/chess/domain/ChessGameTest.java
@@ -1,9 +1,6 @@
 package chess.domain;
 
-import static chess.domain.TestUtils.*;
 import static org.assertj.core.api.Assertions.*;
-
-import java.util.Map;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,10 +13,7 @@ public class ChessGameTest {
     @Test
     @DisplayName("흰 말로 시작하지 않으면 에러를 반환한다")
     void errorTurn_Start() {
-        Board chessBoard = new Board(Map.of(
-                new Square("c3"), WHITE_QUEEN,
-                new Square("d4"), BLACK_QUEEN
-        ));
+        Board chessBoard = new Board(new TestOtherTeamBoard());
         ChessGame chessGame = new ChessGame(chessBoard, Color.WHITE);
         assertThatThrownBy(() -> chessGame.move(new Square("d4"), new Square("e5")))
                 .isInstanceOf(IllegalArgumentException.class)
@@ -29,9 +23,7 @@ public class ChessGameTest {
     @Test
     @DisplayName("흰 말 다음 검은 말 순서로 진행하지 않으면 에러를 반환한다")
     void errorTurn() {
-        Board chessBoard = new Board(Map.of(
-                new Square("c3"), WHITE_QUEEN
-        ));
+        Board chessBoard = new Board(new TestOtherTeamBoard());
         ChessGame chessGame = new ChessGame(chessBoard, Color.WHITE);
         chessGame.move(new Square("c3"), new Square("d4"));
         assertThatThrownBy(() -> chessGame.move(new Square("d4"), new Square("e5")))
